requirements.txt
black
py-solc-x
web3
python-dotenv
